gatewayHostPackage_v2/projects/requirements.txt
docker==4.2.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-SocketIO==4.3.2
paho-mqtt==1.5.0
requests==2.24.0
schedule==0.6.0
Werkzeug==1.0.1
